requirements.txt
Django==4.2.16
djangorestframework
djangorestframework-simplejwt
django-filter
openpyxl
psycopg2
dotenv
confluent_kafka
boto3
pillow
google-generativeai
qrcode
cryptography
selenium
webdriver-manager
beautifulsoup4
celery
redis

